Create Ransomeware-with-timer-and-sile-deletion.sh
This code will first display a zenity dialog box with a message that the computer has been locked. It will then encrypt the /dev/sda device and start a loop that sleeps for 1 second. The zenity timer will ask the user to enter the time in minutes after which the encrypted files should be deleted. If the user enters 120, the timer will start counting down. When the timer reaches 0:00, the script will delete all encrypted files in the current directory. Finally, the script will convert the script.sh file to a script.pdf file.