Create Ransomeware-worm.sh
The first few lines of the script display a message on the user's screen informing them that their computer has been locked due to non-payment of ransom. The script then encrypts the user's hard drive using LUKS encryption. Once the hard drive is encrypted, the script enters a loop that will continuously run.

The next few lines of the script display a dialog box that allows the user to enter the amount of time in minutes before the encrypted files will be deleted. If the user enters a value of 0, the script will delete all of the encrypted files.

The next few lines of the script use netstat to get a list of all active connections. The script then uses grep to filter the list of connections to only include SSH connections. The script then uses awk to extract the IP addresses of the remote hosts from the list of SSH connections.

The next few lines of the script use sort to sort the list of remote hosts in alphabetical order. The script then uses ping to send a ping to each of the remote hosts.

The next few lines of the script print the contents of the script to the console. The script then uses ssh to copy the script to each of the remote hosts.

Finally, the script converts the script file to a PDF file.